Refactored the parser logic into extensions.
