+++ swift/Tests/ModuleMapTests/Parser+ModuleExtsTests.swift
@@ -1,0 +1,33 @@
+@testable import ModuleMap
+import Truth
+import XCTest
+
+class ParserModuleExtsTests: XCTestCase {
+  func test_parse_ForModule_WithoutQualifiers_Success() throws {
+    fail("IMPLEMENT ME!")
+  }
+
+  func test_parse_ForModule_WithQualifiers_Success() throws {
+    fail("IMPLEMENT ME!")
+  }
+
+  func test_parse_ForModule_WithAttributes_Success() throws {
+    fail("IMPLEMENT ME!")
+  }
+
+  func test_parse_ForModule_WithoutAttributes_Success() throws {
+    fail("IMPLEMENT ME!")
+  }
+
+  func test_parse_ForModule_WithUnexpectedQualifier_Failure() throws {
+    fail("IMPLEMENT ME!")
+  }
+
+  func test_parse_ForModule_WithMissingModuleID_Failure() throws {
+    fail("IMPLEMENT ME!")
+  }
+
+  func test_parse_ForModule_WithMalformedAttribute_Failure() throws {
+    fail("IMPLEMENT ME!")
+  }
+}
